# openapi/v1.yml
...
openapi: "3.0.0"
info:
  version: "1.0.0-alpha1"
  description: "API definition for Kleister, manage mod packs for Minecraft"
  title: "Kleister OpenAPI"
  contact:
    name: "Kleister"
    email: "kleister@webhippie.de"
    url: "https://kleister.eu"
  license:
    name: "Apache-2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
servers:
  - url: "https://try.kleister.eu/api/v1"
security:
  - Header: []
  - Bearer: []
  - Basic: []
  - Cookie: []
paths:
...
  /minecraft:
    get:
      summary: "Fetch the available Minecraft versions"
      operationId: "ListMinecrafts"
      tags:
        - "minecraft"
      parameters:
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
      responses:
        "200":
          description: "A collection of minecraft versions"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/minecrafts"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    put:
      summary: "Update the available Minecraft versions"
      operationId: "UpdateMinecraft"
      tags:
        - "minecraft"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "503":
          description: "If remote source is not available"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
  /minecraft/{minecraft_id}/builds:
    get:
      summary: "Fetch the builds attached to a Minecraft version"
      operationId: "ListMinecraftBuilds"
      tags:
        - "minecraft"
      parameters:
        - in: "path"
          name: "minecraft_id"
          description: "A minecraft identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "afcxm2owajfthby9bq1p"
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
        - name: "sort"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "build_name"
            enum:
              - "build_name"
              - "build_public"
              - "pack_slug"
              - "pack_name"
          description: "Sorting column"
          x-example: "name"
        - name: "order"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "asc"
            enum:
              - "asc"
              - "desc"
          description: "Sorting order"
          x-example: "desc"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 100
          description: "Paging limit"
          x-example: 100
        - name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
            default: 0
          description: "Paging offset"
          x-example: 0
      responses:
        "200":
          description: "A collection of attached builds"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/minecraft_builds"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Minecraft or build not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    post:
      summary: "Attach a build to a Minecraft version"
      operationId: "AttachMinecraftToBuild"
      tags:
        - "minecraft"
      parameters:
        - in: "path"
          name: "minecraft_id"
          description: "A minecraft identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "afcxm2owajfthby9bq1p"
      requestBody:
        description: "The build data to attach"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/minecraft_build_params"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Minecraft or build not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "412":
          description: "Build is already attached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    delete:
      summary: "Unlink a build from a Minecraft version"
      operationId: "DeleteMinecraftFromBuild"
      tags:
        - "minecraft"
      parameters:
        - in: "path"
          name: "minecraft_id"
          description: "A minecraft identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "afcxm2owajfthby9bq1p"
      requestBody:
        description: "The build data to unlink"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/minecraft_build_params"
      responses:
        "200":
          description: "Plain success message"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Minecraft or build not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "412":
          description: "Build is not attached"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
...
  /packs/{pack_id}/builds:
    get:
      summary: "Fetch all available builds for a pack"
      operationId: "ListBuilds"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
        - name: "search"
          in: "query"
          required: false
          schema:
            type: "string"
          description: "Search query"
          x-example: "John Doe"
        - name: "sort"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "name"
            enum:
              - "name"
              - "public"
          description: "Sorting column"
          x-example: "name"
        - name: "order"
          in: "query"
          required: false
          schema:
            type: "string"
            default: "asc"
            enum:
              - "asc"
              - "desc"
          description: "Sorting order"
          x-example: "desc"
        - name: "limit"
          in: "query"
          required: false
          schema:
            type: "integer"
          description: "Paging limit"
          x-example: 100
        - name: "offset"
          in: "query"
          required: false
          schema:
            type: "integer"
          description: "Paging offset"
          x-example: 0
      responses:
        "200":
          description: "A collection of builds"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/builds"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"

    post:
      summary: "Create a new build for a pack"
      operationId: "CreateBuild"
      tags:
        - "pack"
      parameters:
        - in: "path"
          name: "pack_id"
          description: "A pack identifier or slug"
          schema:
            type: "string"
          required: true
          x-eample: "edtmbzwztudkdhhfemhd"
      requestBody:
        description: "The build data to create"
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/build"
      responses:
        "200":
          description: "The created build build"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/build"
        "403":
          description: "User is not authorized"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "404":
          description: "Pack not found"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "422":
          description: "Failed to validate request"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        "500":
          description: "Some internal server error"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
        default:
          description: "Some error unrelated to the handler"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/notification"
...
components:
  securitySchemes:
    Header:
      type: apiKey
      in: header
      name: X-API-Key
    Bearer:
      type: http
      scheme: bearer
    Basic:
      type: http
      scheme: basic
    Cookie:
      type: apiKey
      in: header
      name: Cookie
  schemas:
...
    minecrafts:
      title: "Minecrafts"
      description: "Model to represent list of minecrafts"
      type: "object"
      properties:
        total:
          type: integer
          format: int64
        versions:
          type: "array"
          items:
            $ref: "#/components/schemas/minecraft"
    minecraft:
      title: "Minecraft"
      description: "Model to represent minecraft"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    minecraft_builds:
      title: "Minecraft Builds"
      description: "Model to represent minecraft builds"
      type: "object"
      properties:
        minecraft:
          $ref: "#/components/schemas/minecraft"
          readOnly: true
        total:
          type: integer
          format: int64
        builds:
          type: "array"
          items:
            $ref: "#/components/schemas/build"
    minecraft_build_params:
      title: "Minecraft Build Params"
      description: "Model to represent params for minecraft builds"
      type: "object"
      required:
        - "pack"
        - "build"
      properties:
        pack:
          type: "string"
        build:
          type: "string"
...
    forge:
      title: "Forge"
      description: "Model to represent forge"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        minecraft:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    neoforge:
      title: "Neoforge"
      description: "Model to represent neoforge"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    quilt:
      title: "Quilt"
      description: "Model to represent quilt"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    fabric:
      title: "Fabric"
      description: "Model to represent fabric"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    pack:
      title: "Pack"
      description: "Model to represent pack"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        icon:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_icon"
        logo:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_logo"
        back:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack_back"
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        website:
          type: "string"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    pack_icon:
      title: "Pack Icon"
      description: "Model to represent pack icon"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        content_type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        md5:
          type: "string"
          x-omitempty: true
          x-nullable: true
        path:
          type: "string"
          x-omitempty: true
          x-nullable: true
        url:
          type: "string"
          x-omitempty: true
          x-nullable: true
        upload:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    pack_logo:
      title: "Pack Logo"
      description: "Model to represent pack logo"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        content_type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        md5:
          type: "string"
          x-omitempty: true
          x-nullable: true
        path:
          type: "string"
          x-omitempty: true
          x-nullable: true
        url:
          type: "string"
          x-omitempty: true
          x-nullable: true
        upload:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    pack_back:
      title: "Pack Background"
      description: "Model to represent pack background"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        content_type:
          type: "string"
          x-omitempty: true
          x-nullable: true
        md5:
          type: "string"
          x-omitempty: true
          x-nullable: true
        path:
          type: "string"
          x-omitempty: true
          x-nullable: true
        url:
          type: "string"
          x-omitempty: true
          x-nullable: true
        upload:
          type: "string"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
    builds:
      title: "Builds"
      description: "Model to represent list of builds"
      type: "object"
      properties:
        total:
          type: integer
          format: int64
        pack:
          readOnly: true
          $ref: "#/components/schemas/pack"
        builds:
          type: "array"
          items:
            $ref: "#/components/schemas/build"
    build:
      title: "Build"
      description: "Model to represent build"
      type: "object"
      properties:
        id:
          type: "string"
          readOnly: true
        pack:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/pack"
        minecraft_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        minecraft:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/minecraft"
        forge_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        forge:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/forge"
        neoforge_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        neoforge:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/neoforge"
        quilt_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        quilt:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/quilt"
        fabric_id:
          type: "string"
          x-omitempty: true
          x-nullable: true
        fabric:
          x-omitempty: true
          x-nullable: true
          readOnly: true
          $ref: "#/components/schemas/fabric"
        slug:
          type: "string"
          x-omitempty: true
          x-nullable: true
        name:
          type: "string"
          x-omitempty: true
          x-nullable: true
        java:
          type: "string"
          x-omitempty: true
          x-nullable: true
        memory:
          type: "string"
          x-omitempty: true
          x-nullable: true
        latest:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        recommended:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        public:
          type: "boolean"
          x-omitempty: true
          x-nullable: true
        created_at:
          type: "string"
          format: "date-time"
          readOnly: true
        updated_at:
          type: "string"
          format: "date-time"
          readOnly: true
...
    notification:
      title: "Notification"
      description: "Generic response for errors and validations"
      type: "object"
      properties:
        status:
          type: "integer"
          x-omitempty: true
        message:
          type: "string"
          x-omitempty: true
        errors:
          type: "array"
          items:
            $ref: "#/components/schemas/validation"
    validation:
      title: "Validation"
      description: "General structure to show validation errors"
      type: "object"
      properties:
        field:
          type: "string"
        message:
          type: "string"
...